SLAProject/requirements.txt
opencv-python-headless
streamlit
numpy
cvzone
torch
torchvision
Pillow
